ws-160 split the integration test files
